rag + db version/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.3

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

sentence-transformers
huggingface-hub
torch
transformers==4.30.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2

# Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.0
selenium